Organise Library
Create Objet Constructor for the books and add the book object
in library each time a book is created.
Display the book on the page by wrapping the infos in a card who is wraped
in a div.container
Add some basics styles to the page to visualize better.
Pop a dialog form on the page to add infos for a book the user want to add